+++ client/config/default.json
@@ -1,0 +1,6 @@
+{
+    "dynamodbURI": {
+        "region": "us-east-1",
+        "endpoint": "https://dynamodb.us-east-1.amazonaws.com"
+    }
+}
