+++ client/src/api/cause.js
@@ -6,10 +6,9 @@
 // Alternatively use async/await syntax:
 // `await org.getOrgById( ... );`
 
-'use strict';
-const api = require('../client/src/utils/api');
+const api = require('../utils/api');
 const AWS = require('aws-sdk');
-const { docClient } = require('../config/db');
+const connectDB = require('../db');
 connectDB();
 const docClient = new AWS.DynamoDB.DocumentClient();
 
@@ -31,7 +30,8 @@ const getAll = async () => {
         const params = {
             TableName: 'cause'
         };
-        return await docClient.scan(params).promise();
+        const res = await docClient.scan(params).promise();
+        return res.Items;
     } catch (err) { throw err }
 };
