--- middleware/auth.js
@@ -1,1 +1,0 @@
-// TODO auth middleware
