--- models/Post.js
@@ -1,1 +1,0 @@
-// TODO make `Post` schema.
